Dead end for `Catchable` protocol?
In order to have a protocol called Catchable we need to conform a GenericNode to Catchable multiple times: once for if it is a Result and once for if it is a Promise. But this is not allowed. The alternatives: make subclasses rather than have everything be typealias, or have the body of `catch` switch on the type of When